change dimension from int to double

+++ src/main/java/it/unibo/pensilina14/bullet/ballet/common/Dimension2D.java
@@ -3,22 +3,23 @@ package it.unibo.pensilina14.bullet.ballet.common;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 
 public interface Dimension2D {
-  /**
-   * 
-   * @return object's dimensions
-   */
-  ImmutablePair<Integer, Integer> getSize();
-  
-  /**
-   * 
-   * @return object's height
-   */
-  int getHeight();
-  
-  /**
-   * 
-   * @return object's width
-   */
-  int getWidth();
+    
+    /**
+     * 
+     * @return object's dimensions
+     */
+    ImmutablePair<Double, Double> getSize();
+    
+    /**
+     * 
+     * @return object's height
+     */
+    double getHeight();
+    
+    /**
+     * 
+     * @return object's width
+     */
+    double getWidth();
     
 }
